Changed the entire algorithm to compliment every case
The algorithm has been changed to check and see if there is an available space that also satisfies the given condition (e.x. closer to fish, away from whale, etc.) Also added the random penguin movement. Scenarios where the penguin could be in the corner or edge have also been accounted for. It uses several functions that still need to be implemented in other classes, but most are just accessor and mutator functions. 

The key, however, is to implement the filled() function which returns true if the space has a whale, is empty, or is out of the grid. The algorithm uses this heavily, and must be implemented in the Sea class soon.